requirements.txt
streamlit>=1.26
opencv-python>=4.8
numpy>=1.24
Pillow>=10.0
